+++ tests/integration/test_store.py
@@ -284,12 +284,15 @@ def make_error(event):
 
 
 @pytest.mark.parametrize("event_type", ["default", "transaction"])
+@pytest.mark.parametrize("use_arroyo", [False, True], ids=["rdkafka", "arroyo"])
 def test_processing(
     mini_sentry,
     relay_with_processing,
     events_consumer,
     transactions_consumer,
+    processing_config,
     event_type,
+    use_arroyo,
 ):
     """
     Test that relay normalizes messages when processing is enabled and sends them via Kafka queues
@@ -300,7 +303,8 @@ def test_processing(
     else:
         events_consumer = transactions_consumer()
 
-    relay = relay_with_processing()
+    options = processing_config({"processing": {"use_arroyo": use_arroyo}})
+    relay = relay_with_processing(options=options)
     project_id = 42
     mini_sentry.add_full_project_config(42)